Substantial changing in order to process payload

+++ anisotropy_utils.py
@@ -94,7 +94,7 @@ def anisotropy_rotation_one_axis(matrix, theta, phi, beta):
         axis=-2,
     )
 
-    total_rotation = rotation_z @ rotation_y @ rotation_x
+    total_rotation = tf.cast(rotation_z @ rotation_y @ rotation_x, dtype=tf.complex64)
 
     matrix = matrix[:, tf.newaxis, :, :]
     total_rotation = total_rotation[tf.newaxis, :, :, :]
